+++ config/nvim/lua/README.md
@@ -1,0 +1,2 @@
+# Neovim lua
+All files in here can be directly `require`'d from neovim lua
