--- config/nvim/lua/config/cmp.lua
@@ -1,13 +1,0 @@
--- local conf = {}
-
--- local function has_words_before()
--- 	local line, col = unpack(vim.api.nvim_win_get_cursor(0))
--- 	return col ~= 0 and vim.api.nvim_buf_get_lines(0, line - 1, line, true)[1]:sub(col, col):match("%s") == nil
--- end
-
--- local function setup_cmp_highlighting()
--- end
-
--- function conf.setup() end
-
--- return conf
