+++ config/nvim/lua/plugins/easy_align.lua
@@ -1,0 +1,14 @@
+return {
+	"junegunn/vim-easy-align",
+	init = function()
+		vim.keymap.set("x", "ga", function()
+			require("lazy").load("vim-easy-align")
+			return "<Plug>(EasyAlign)"
+		end, { noremap = false, expr = true })
+
+		vim.keymap.set("n", "ga", function()
+			require("lazy").load("vim-easy-align")
+			return "<Plug>(EasyAlign)"
+		end, { noremap = false, expr = true })
+	end,
+}
