+++ config/nvim/lua/plugins/gitsigns.lua
@@ -111,7 +111,7 @@ end
 return {
 	"lewis6991/gitsigns.nvim",
 	event = "BufReadPost",
-	requires = {
+	dependencies = {
 		"nvim-lua/plenary.nvim",
 	},
 	config = function()
