+++ config/nvim/lua/plugins/init.lua
@@ -11,6 +11,25 @@ return {
 	{ "mechatroner/rainbow_csv", ft = "csv" },
 	{ "tpope/vim-vinegar", keys = { "-" }, cmd = { "Explore" } },
 	{
+		"rcarriga/nvim-notify",
+		config = function()
+			require("notify").setup({
+				stages = "slide",
+				timeout = 3000,
+				-- Minimum width for notification windows
+				minimum_width = 30,
+				icons = {
+					ERROR = "",
+					WARN = "",
+					INFO = "",
+					DEBUG = "",
+					TRACE = "✎",
+				},
+			})
+			vim.notify = require("notify")
+		end,
+	},
+	{
 		"folke/zen-mode.nvim",
 		cmd = {
 			"ZenMode",
@@ -24,7 +43,6 @@ return {
 			})
 		end,
 	},
-
 	{
 		"williamboman/mason.nvim",
 		-- requires = { "williamboman/mason-lspconfig.nvim" },
@@ -40,19 +58,34 @@ return {
 			require("config.luasnip")
 		end,
 	},
+	{
+		"nicwest/vim-camelsnek",
+		cmd = { "Snek", "Pascal", "Camel", "Kebab" },
+		init = function()
+			vim.g.camelsnek_alternative_camel_commands = 1
 
+			vim.keymap.set("n", "crs", ":Snek<CR>", { silent = true, desc = "snake_case" })
+			vim.keymap.set("x", "crs", ":Snek<CR>", { silent = true, desc = "snake_case" })
+
+			vim.keymap.set("n", "crp", ":Pascal<CR>", { silent = true, desc = "PascalCase" })
+			vim.keymap.set("x", "crp", ":Pascal<CR>", { silent = true, desc = "PascalCase" })
+
+			vim.keymap.set("n", "crc", ":Camel<CR>", { silent = true, desc = "camelCase" })
+			vim.keymap.set("x", "crc", ":Camel<CR>", { silent = true, desc = "camel_case" })
+
+			vim.keymap.set("n", "crk", ":Kebab<CR>", { silent = true, desc = "kebab-case" })
+			vim.keymap.set("x", "crk", ":Kebab<CR>", { silent = true, desc = "kebab-case" })
+		end,
+	},
 	{
 		"windwp/nvim-autopairs",
 		event = "InsertEnter",
 		config = function()
-			require("config.general").autopairs()
-		end,
-	},
-	{
-		"mhinz/vim-startify",
-		branch = "center",
-		config = function()
-			require("config.startify").setup()
+			require("nvim-autopairs").setup({
+				fast_wrap = {},
+				check_ts = true,
+				disable_in_macro = true,
+			})
 		end,
 	},
 }
