+++ config/nvim/lua/plugins/startify.lua
@@ -1,0 +1,382 @@
+local conf = {}
+
+local function center(list)
+	return vim.fn["startify#center"](list)
+end
+
+local running_man = {
+	[[  \                           /      ]],
+	[[   \                         /       ]],
+	[[    \                       /        ]],
+	[[     ]                     [    ,'|  ]],
+	[[     ]               &&&&  [   /  |  ]],
+	[[     ]___           &&&&&&_[ ,'   |  ]],
+	[[     ]  ]\          &&&/[  [ |:   |  ]],
+	[[     ]  ] \          &/ [  [ |:   |  ]],
+	[[     ]  ]  ]         [  [  [ |:   |  ]],
+	[[     ]  ]  ]__     __[  [  [ |:   |  ]],
+	[[     ]  ]  ] ]\ _ /[ [  [  [ |:   |  ]],
+	[[     ]  ]  ] ] (#) [ [  [  [ :===='  ]],
+	[[     ]  ]  ]_].nHn.[_[  [  [         ]],
+	[[     ]  ]  ]  HHHHH. [  [  [         ]],
+	[[     ]  ] /   `HH("N  \ [  [         ]],
+	[[     ]__]/     HHH  "  \[__[         ]],
+	[[     ]         NNN         [         ]],
+	[[     ]         N/"         [         ]],
+	[[     ]         N H         [         ]],
+	[[    /          N            \        ]],
+	[[   /           q,            \       ]],
+	[[  /                           \      ]],
+}
+
+local generate_trees = function()
+	local smallest = {
+		[[            ]],
+		[[      /&*/  ]],
+		[[  &/&/*/ &  ]],
+		[[  &/ \}/ {_&]],
+		[[*@\_{*\    *]],
+		[[    {}      ]],
+		[[    {{}     ]],
+		[[=-~{ .-^-\  ]],
+		[[    `}      ]],
+		[[      \     ]],
+	}
+
+	local baby = {
+		[[        *      ]],
+		[[  &     &&&    ]],
+		[[&&|, &&|/*@&   ]],
+		[[  &/&&*/ &/_&  ]],
+		[[  && \\/ {_&  @]],
+		[[*@\_{*&//      ]],
+		[[    `, \{      ]],
+		[[    } }        ]],
+		[[    {{}        ]],
+		[[-=-~{ .-^-\    ]],
+		[[    `}         ]],
+		[[      \        ]],
+	}
+
+	local teen = {
+		[[         &&           ]],
+		[[       & \/&          ]],
+		[[     && \/& &&&       ]],
+		[[    &&|,/  |/  &&     ]],
+		[[ *&_ &&/___/  /_&  _  ]],
+		[[&_\\    \\/ { &|__/ * ]],
+		[[   \\___{*&/ /       _]],
+		[[       `, \{____&__&/*]],
+		[[        } }/          ]],
+		[[        }{            ]],
+		[[       {{}            ]],
+		[[ ,--=-~{ .-^-\_,_     ]],
+		[[       `}             ]],
+		[[        {             ]],
+	}
+
+	local adult = {
+		[[           &&                 ]],
+		[[         &&&&&                ]],
+		[[ ,  &  &&&\/& &&&             ]],
+		[[&-\* &&&|,/ *|/& &&           ]],
+		[[ . *&_ &&/   /, /_&  &&       ]],
+		[[  &_\\    \\  { &|__&_&/_&    ]],
+		[[ ^   \\_&_{*&/ /          &&& ]],
+		[[         `, \{__&_&__&_&_/_&& ]],
+		[[          } }{ * @  &\\'      ]],
+		[[          }{{         \'_&__& ]],
+		[[         {}{           \ `&\&&]],
+		[[         {{}             '&@  ]],
+		[[   ,--=-~{ .-^-\_,_           ]],
+		[[         `}                   ]],
+		[[          {                   ]],
+	}
+
+	local middle_aged = {
+		[[            &&                 ]],
+		[[          &&&&&                ]],
+		[[        &&&\/& &&&             ]],
+		[[       &&|,/  |/& &&           ]],
+		[[    *&_ &&/   /  /_&  &&       ]],
+		[[   &_\\    \\  { &|__&_&/_&    ]],
+		[[      \\_&_{*&/ /          &&& ]],
+		[[   ,      `, \{__&_&__&_&_/_&& ]],
+		[[  &        } }{      &\\'      ]],
+		[[           }{{         \'_&__& ]],
+		[[          {}{           \ `&\&&]],
+		[[          {{}             '&&  ]],
+		[[    ,--=-~{ .-^-\_,_           ]],
+		[[&,        `}                   ]],
+		[[  .        {                   ]],
+	}
+
+	local old = {
+		[[            _                 ]],
+		[[             \                ]],
+		[[          &\/{  _/            ]],
+		[[       &_|,/} {/& &&          ]],
+		[[    *&_  \/ | /  /_&  &&      ]],
+		[[   &_\\   \ \  { &|__&_&/_&   ]],
+		[[      \\_\_{*&/ /          &&&]],
+		[[   ,      `, \{__&_&__&_&_/_&&]],
+		[[  &        } }{      &\\'     ]],
+		[[           }{{         \'____&]],
+		[[          {}{           \ ` \ ]],
+		[[          {{}           *     ]],
+		[[ . &,-.&-~{ .-^-*_,_          ]],
+		[[&,   @    `}   .& .      _@&& ]],
+		[[  .        {     .  .    \**  ]],
+	}
+
+	local older = {
+		[[                              ]],
+		[[                              ]],
+		[[          &\/{  _/      .&    ]],
+		[[       &_  /} {/  _&          ]],
+		[[    *__  \/ | /  /    &       ]],
+		[[    _\\   \ \{ { /\__/        ]],
+		[[   /  \__\_{*_/ /     \     .*]],
+		[[   |   }  `,/\{__&__/__       ]],
+		[[           } }{ }     \       ]],
+		[[       .   }{{   *    ,       ]],
+		[[   & .    {}{                 ]],
+		[[          {{}      .  ,       ]],
+		[[ . _,-._-~{ .-^-*_,_   *      ]],
+		[[ ,  _/   /_}\ \_    \_   _    ]],
+		[[  /  .     {      . .     \*  ]],
+	}
+
+	local old_boxed = {
+		[[._____. ._____. .____________________. ._____. ._____.]],
+		[[| ._. | | ._. | | .________________. | | ._. | | ._. |]],
+		[[| !_| |_|_|_! | | !________________! | | !_| |_|_|_! |]],
+		[[!___| |_______! !____________________! !___| |_______!]],
+		[[.___|_|_| |________________________________|_|_| |___.]],
+		[[| ._____| |____________________________________| |_. |]],
+		[[| !_! | | |                                | | ! !_! |]],
+		[[!_____! | |             _                  | | !_____!]],
+		[[._____. | |              \                 | | ._____.]],
+		[[| ._. | | |           &\/{  _/             | | | ._. |]],
+		[[| | | | | |        &_|,/} {/& &&           | | | | | |]],
+		[[| | | | | |     *&_  \/ | /  /_&  &&       | | | | | |]],
+		[[| | | | | |    &_\\   \ \  { &|__&_&/_&    | | | | | |]],
+		[[| | | | | |       \\_\_{*&/ /          &&& | | | | | |]],
+		[[| | | | | |    ,      `, \{__&_&__&_&_/_&& | | | | | |]],
+		[[| | | | | |   &        } }{      &\\'      | | | | | |]],
+		[[| | | | | |            }{{         \'____& | | | | | |]],
+		[[| | | | | |           {}{           \ ` \  | | | | | |]],
+		[[| !_! | | |           {{}           *      | | ! !_! |]],
+		[[!_____! | |  . &,-.&-~{ .-^-*_,_           | | !_____!]],
+		[[._____. | | &,   @    `}   .& .      _@&&  | | ._____.]],
+		[[| ._. | | |   .        {     .  .    \**   | | | ._. |]],
+		[[| !_| |_|_|________________________________| |_|_|_! |]],
+		[[!___| |____________________________________| |_______!]],
+		[[.___|_|_| |___. .____________________. .___|_|_| |___.]],
+		[[| ._____| |_. | | .________________. | | ._____| |_. |]],
+		[[| !_! | | !_! | | !________________! | | !_! | | !_! |]],
+		[[!_____! !_____! !____________________! !_____! !_____!]],
+	}
+
+	local middle_aged_boxed = {
+		[[._____. ._____. ._____________________. ._____. ._____.]],
+		[[| ._. | | ._. | | ._________________. | | ._. | | ._. |]],
+		[[| !_| |_|_|_! | | !_________________! | | !_| |_|_|_! |]],
+		[[!___| |_______! !_____________________! !___| |_______!]],
+		[[.___|_|_| |_________________________________|_|_| |___.]],
+		[[| ._____| |_____________________________________| |_. |]],
+		[[| !_! | | |                                 | | ! !_! |]],
+		[[!_____! | |             &&                  | | !_____!]],
+		[[._____. | |           &&&&&                 | | ._____.]],
+		[[| ._. | | |         &&&\/& &&&              | | | ._. |]],
+		[[| | | | | |        &&|,/  |/& &&            | | | | | |]],
+		[[| | | | | |     *&_ &&/   /  /_&  &&        | | | | | |]],
+		[[| | | | | |    &_\\    \\  { &|__&_&/_&     | | | | | |]],
+		[[| | | | | |       \\_&_{*&/ /          &&&  | | | | | |]],
+		[[| | | | | |    ,      `, \{__&_&__&_&_/_&&  | | | | | |]],
+		[[| | | | | |   &        } }{      &\\'       | | | | | |]],
+		[[| | | | | |            }{{         \'_&__&  | | | | | |]],
+		[[| | | | | |           {}{           \ `&\&& | | | | | |]],
+		[[| !_! | | |           {{}             '&&   | | ! !_! |]],
+		[[!_____! | |     ,--=-~{ .-^-\_,_            | | !_____!]],
+		[[._____. | | &,        `}                    | | ._____.]],
+		[[| ._. | | |   .        {                    | | | ._. |]],
+		[[| !_| |_|_|_________________________________| |_|_|_! |]],
+		[[!___| |_____________________________________| |_______!]],
+		[[.___|_|_| |___. ._____________________. .___|_|_| |___.]],
+		[[| ._____| |_. | | ._________________. | | ._____| |_. |]],
+		[[| !_! | | !_! | | !_________________! | | !_! | | !_! |]],
+		[[!_____! !_____! !_____________________! !_____! !_____!]],
+	}
+
+	local adult_boxed = {
+		[[._____. ._____. .____________________. ._____. ._____.]],
+		[[| ._. | | ._. | | .________________. | | ._. | | ._. |]],
+		[[| !_| |_|_|_! | | !________________! | | !_| |_|_|_! |]],
+		[[!___| |_______! !____________________! !___| |_______!]],
+		[[.___|_|_| |________________________________|_|_| |___.]],
+		[[| ._____| |____________________________________| |_. |]],
+		[[| !_! | | |                                | | ! !_! |]],
+		[[!_____! | |            &&                  | | !_____!]],
+		[[._____. | |          &&&&&                 | | ._____.]],
+		[[| ._. | | |  ,  &  &&&\/& &&&              | | | ._. |]],
+		[[| | | | | | &-\* &&&|,/ *|/& &&            | | | | | |]],
+		[[| | | | | |  . *&_ &&/   /, /_&  &&        | | | | | |]],
+		[[| | | | | |   &_\\    \\  { &|__&_&/_&     | | | | | |]],
+		[[| | | | | |  ^   \\_&_{*&/ /          &&&  | | | | | |]],
+		[[| | | | | |          `, \{__&_&__&_&_/_&&  | | | | | |]],
+		[[| | | | | |           } }{ * @  &\\'       | | | | | |]],
+		[[| | | | | |           }{{         \'_&__&  | | | | | |]],
+		[[| | | | | |          {}{           \ `&\&& | | | | | |]],
+		[[| !_! | | |          {{}             '&@   | | ! !_! |]],
+		[[!_____! | |    ,--=-~{ .-^-\_,_            | | !_____!]],
+		[[._____. | |          `}                    | | ._____.]],
+		[[| ._. | | |           {                    | | | ._. |]],
+		[[| !_| |_|_|________________________________| |_|_|_! |]],
+		[[!___| |____________________________________| |_______!]],
+		[[.___|_|_| |___. .____________________. .___|_|_| |___.]],
+		[[| ._____| |_. | | .________________. | | ._____| |_. |]],
+		[[| !_! | | !_! | | !________________! | | !_! | | !_! |]],
+		[[!_____! !_____! !____________________! !_____! !_____!]],
+		[[                                                      ]],
+	}
+
+	local teen_boxed = {
+		[[       /\          /\          /\       ]],
+		[[    /\//\\/\    /\//\\/\    /\//\\/\    ]],
+		[[ /\//\\\///\\/\//\\\///\\/\//\\\///\\/\ ]],
+		[[//\\\//\/\\///\\\//\/\\///\\\//\/\\///\\]],
+		[[\\//\/                            \/\\//]],
+		[[ \/                &&                \/ ]],
+		[[ /\              & \/&               /\ ]],
+		[[//\\           && \/& &&&           //\\]],
+		[[\\//         &&|,/  |/  &&          \\//]],
+		[[ \/       *&_ &&/___/  /_&  _        \/ ]],
+		[[ /\      &_\\    \\/ { &|__/ *       /\ ]],
+		[[//\\      \\___{*&/ /       _       //\\]],
+		[[\\//        `, \{____&__&/*         \\//]],
+		[[ \/               } }/               \/ ]],
+		[[ /\                }{                /\ ]],
+		[[//\\              {{}               //\\]],
+		[[\\//        ,--=-~{ .-^-\_,_        \\//]],
+		[[ \/                `}                \/ ]],
+		[[ /\                {                 /\ ]],
+		[[//\\/\                            /\//\\]],
+		[[\\///\\/\//\\\///\\/\//\\\///\\/\//\\\//]],
+		[[ \/\\///\\\//\/\\///\\\//\/\\///\\\//\/ ]],
+		[[    \/\\//\/    \/\\//\/    \/\\//\/    ]],
+		[[       \/          \/          \/       ]],
+	}
+
+	local baby_boxed = {
+		[[        <\/><\/><\/>        ]],
+		[[    <\/></\></\></\><\/>    ]],
+		[[    </\>            </\>    ]],
+		[[ <\/>        *         <\/> ]],
+		[[ </\>    &     &&&     </\> ]],
+		[[<\/>    &&|, &&|/*@&    <\/>]],
+		[[</\>    &/&&*/ &/_&     </\>]],
+		[[<\/>   && \\/ {_&  @    <\/>]],
+		[[</\>     *@\_{*&//      </\>]],
+		[[<\/>       `, \{        <\/>]],
+		[[</\>        } }         </\>]],
+		[[<\/>        {{}         <\/>]],
+		[[</\>    -=-~{ .-^-\     </\>]],
+		[[ <\/>        `}        <\/> ]],
+		[[ </\>        \         </\> ]],
+		[[    <\/>            <\/>    ]],
+		[[    </\><\/><\/><\/></\>    ]],
+		[[        </\></\></\>        ]],
+	}
+
+	local smallest_boxed = {
+		[[ .+"+.+"+.+"+.+"+. ]],
+		[[(        &        )]],
+		[[ )     &&|/      ( ]],
+		[[(    &/&&*/ &     )]],
+		[[ )  && \}/ {_&   ( ]],
+		[[(     *@\_{*&     )]],
+		[[ )      {}       ( ]],
+		[[(       {{}       )]],
+		[[ )  =-~{ .-^-\   ( ]],
+		[[(       `}        )]],
+		[[ )       \       ( ]],
+		[[(                 )]],
+		[[ "+.+"+.+"+.+"+.+" ]],
+	}
+
+	return {
+		boxed = {
+			smallest_boxed,
+			baby_boxed,
+			teen_boxed,
+			adult_boxed,
+			middle_aged_boxed,
+			old_boxed,
+		},
+		unboxed = {
+			smallest,
+			baby,
+			teen,
+			adult,
+			middle_aged,
+			old,
+		},
+	}
+end
+
+return {
+	"mhinz/vim-startify",
+	branch = "center",
+	config = function()
+		local trees = generate_trees()
+		vim.g.startify_center = 58
+		vim.g.startify_commands = {
+			{ p = { "Files", ":Telescope find_files" } },
+			{ l = { "Lazy", ":Lazy" } },
+			{ d = { "Open dotfiles", ":!dmux ~/yakko_wakko" } },
+			{ D = { "Dmux", ":!dmux" } },
+		}
+
+		vim.g.startify_lists = {
+			{ type = "commands", header = center({ "めいれい" }) },
+			-- { type = "dir", header = center({ "MRU " .. vim.fn.getcwd() }) },
+		}
+
+		vim.g.sttartify_change_to_dir = 0
+		vim.g.startify_change_to_vcs_root = 1
+		local default = adult_boxed
+		local win_height = vim.fn.winheight("%")
+		local win_width = vim.fn.winwidth("%")
+
+		local boxes
+		if (win_width > 120) and (win_height > 50) then
+			boxes = trees["boxed"]
+		else
+			boxes = trees["unboxed"]
+		end
+
+		local day = tonumber(vim.fn.strftime("%H"))
+		local interval = 24 / #boxes
+		local index = math.floor(day / interval)
+		local ascii = boxes[index]
+
+		if not ascii then
+			ascii = default
+		end
+
+		vim.g.startify_custom_header = center(ascii)
+
+		-- local startify_group = vim.api.nvim_create_augroup("startify", { clear = true })
+		-- vim.api.nvim_create_autocmd({ "User" }, {
+		-- 	group = startify_group,
+		-- 	pattern = "StartifyReady",
+		-- 	callback = function()
+		-- 		vim.keymap.set("n", "-", function()
+		-- 			vim.cmd("bwipe")
+		-- 			vim.cmd("Dirvish")
+		-- 		end, { silent = true, buffer = true })
+		-- 	end,
+		-- })
+	end,
+}
