+++ config/nvim/lua/settings.lua
@@ -89,8 +89,8 @@ local disabled_built_ins = {
 	"logipat",
 	-- "netrw",
 	-- "netrwPlugin",
-	"netrwSettings",
-	"netrwFileHandlers",
+	-- "netrwSettings",
+	-- "netrwFileHandlers",
 	-- "matchit",
 	-- "matchparen",
 	"tar",
